# openapi/qe-report-api-v1.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "qe-report API",
    "version": "1",
    "description": "Transport API version 1 of qe-report. It is independent of the protocol version (0.3), the database schema version, the query-index version, and the package version. Every authenticated operation acts on the project its API key belongs to; no request names a project. Bearer API keys must only travel over HTTPS; TLS is expected at the deployment or reverse-proxy boundary.",
    "license": {
      "name": "Apache-2.0",
      "identifier": "Apache-2.0"
    }
  },
  "paths": {
...
    "/v1/runs": {
      "post": {
        "operationId": "ingestRun",
        "summary": "Upload one complete run as multipart protocol source and attachment bytes.",
        "security": [
          {
            "apiKey": []
          }
        ],
        "x-qe-report-scope": "runs:write",
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "description": "The project is the API key’s; no part names one. Event streams are stored exactly as sent. Unknown part names are refused.",
                "properties": {
                  "expiresAt": {
                    "type": "string",
                    "format": "date-time",
                    "description": "Required, exactly once, as text: an RFC 3339 instant with an explicit offset after which retention may delete the run. A past instant is valid."
                  },
                  "events": {
                    "type": "array",
                    "minItems": 1,
                    "items": {
                      "type": "string",
                      "format": "binary"
                    },
                    "description": "One or more NDJSON protocol event streams, as file parts."
                  },
                  "attachment": {
                    "type": "array",
                    "items": {
                      "type": "string",
                      "format": "binary"
                    },
                    "description": "Zero or more attachment byte streams, as file parts, identified by the SHA-256 of their bytes; filenames and media types are ignored."
                  }
                },
                "required": [
                  "expiresAt",
                  "events"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "The same run was already archived; derived state may have been repaired.",
            "headers": {
              "X-Request-Id": {
                "description": "The id this server gave the request; quote it when reporting a problem.",
                "schema": {
                  "type": "string",
                  "format": "uuid"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IngestionResult"
                }
              }
            }
          },
          "201": {
            "description": "Archived now.",
            "headers": {
              "X-Request-Id": {
                "description": "The id this server gave the request; quote it when reporting a problem.",
                "schema": {
                  "type": "string",
                  "format": "uuid"
                }
              },
              "Location": {
                "description": "The run, by its runRef.",
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IngestionResult"
                }
              }
            }
          },
          "400": {
            "description": "The request itself is malformed: its multipart framing or parts.",
            "headers": {
              "X-Request-Id": {
                "description": "The id this server gave the request; quote it when reporting a problem.",
                "schema": {
                  "type": "string",
                  "format": "uuid"
                }
              }
            },
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "401": {
            "description": "No valid API key. Unknown, malformed, expired, and revoked keys are not told apart.",
            "headers": {
              "X-Request-Id": {
                "description": "The id this server gave the request; quote it when reporting a problem.",
                "schema": {
                  "type": "string",
                  "format": "uuid"
                }
              }
            },
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "403": {
            "description": "The key is valid but lacks the scope this operation needs.",
            "headers": {
              "X-Request-Id": {
                "description": "The id this server gave the request; quote it when reporting a problem.",
                "schema": {
                  "type": "string",
                  "format": "uuid"
                }
              }
            },
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "409": {
            "description": "Another run is archived under this run id in this project.",
            "headers": {
              "X-Request-Id": {
                "description": "The id this server gave the request; quote it when reporting a problem.",
                "schema": {
                  "type": "string",
                  "format": "uuid"
                }
              }
            },
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "413": {
            "description": "A transport limit was exceeded; nothing was written.",
            "headers": {
              "X-Request-Id": {
                "description": "The id this server gave the request; quote it when reporting a problem.",
                "schema": {
                  "type": "string",
                  "format": "uuid"
                }
              }
            },
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "415": {
            "description": "The body is not multipart/form-data.",
            "headers": {
              "X-Request-Id": {
                "description": "The id this server gave the request; quote it when reporting a problem.",
                "schema": {
                  "type": "string",
                  "format": "uuid"
                }
              }
            },
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "422": {
            "description": "The protocol source is invalid, incomplete, or empty; nothing was written.",
            "headers": {
              "X-Request-Id": {
                "description": "The id this server gave the request; quote it when reporting a problem.",
                "schema": {
                  "type": "string",
                  "format": "uuid"
                }
              }
            },
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      },
      "get": {
        "operationId": "listRuns",
        "summary": "The project’s runs, newest archived first, in keyset pages.",
        "security": [
          {
            "apiKey": []
          }
        ],
        "x-qe-report-scope": "runs:read",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 1000
            }
          },
          {
            "name": "cursor",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 2048,
              "description": "Opaque. Pass back exactly as received, to the same listing, with the same key."
            }
          }
        ],
        "responses": {
          "200": {
            "description": "One page.",
            "headers": {
              "X-Request-Id": {
                "description": "The id this server gave the request; quote it when reporting a problem.",
                "schema": {
                  "type": "string",
                  "format": "uuid"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunPage"
                }
              }
            }
          },
          "400": {
            "description": "A malformed cursor or limit.",
            "headers": {
              "X-Request-Id": {
                "description": "The id this server gave the request; quote it when reporting a problem.",
                "schema": {
                  "type": "string",
                  "format": "uuid"
                }
              }
            },
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "401": {
            "description": "No valid API key. Unknown, malformed, expired, and revoked keys are not told apart.",
            "headers": {
              "X-Request-Id": {
                "description": "The id this server gave the request; quote it when reporting a problem.",
                "schema": {
                  "type": "string",
                  "format": "uuid"
                }
              }
            },
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "403": {
            "description": "The key is valid but lacks the scope this operation needs.",
            "headers": {
              "X-Request-Id": {
                "description": "The id this server gave the request; quote it when reporting a problem.",
                "schema": {
                  "type": "string",
                  "format": "uuid"
                }
              }
            },
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "503": {
            "description": "The project’s query index is incomplete; nothing partial is answered.",
            "headers": {
              "X-Request-Id": {
                "description": "The id this server gave the request; quote it when reporting a problem.",
                "schema": {
                  "type": "string",
                  "format": "uuid"
                }
              }
            },
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/runs/{runRef}/attachments/{sha256}": {
      "get": {
        "operationId": "getAttachment",
        "summary": "The bytes of an attachment the run references, streamed.",
        "security": [
          {
            "apiKey": []
          }
        ],
        "x-qe-report-scope": "runs:read",
        "parameters": [
          {
            "name": "runRef",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^[A-Za-z0-9_-]{2,172}$",
              "description": "Unpadded base64url of the UTF-8 run id: a transport locator only, never an identity."
            }
          },
          {
            "name": "sha256",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^[0-9a-f]{64}$"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The bytes, always as application/octet-stream.",
            "headers": {
              "X-Request-Id": {
                "description": "The id this server gave the request; quote it when reporting a problem.",
                "schema": {
                  "type": "string",
                  "format": "uuid"
                }
              },
              "X-Content-Type-Options": {
                "description": "Always nosniff.",
                "schema": {
                  "type": "string",
                  "enum": [
                    "nosniff"
                  ]
                }
              }
            },
            "content": {
              "application/octet-stream": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          },
          "401": {
            "description": "No valid API key. Unknown, malformed, expired, and revoked keys are not told apart.",
            "headers": {
              "X-Request-Id": {
                "description": "The id this server gave the request; quote it when reporting a problem.",
                "schema": {
                  "type": "string",
                  "format": "uuid"
                }
              }
            },
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "403": {
            "description": "The key is valid but lacks the scope this operation needs.",
            "headers": {
              "X-Request-Id": {
                "description": "The id this server gave the request; quote it when reporting a problem.",
                "schema": {
                  "type": "string",
                  "format": "uuid"
                }
              }
            },
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "404": {
            "description": "No such run in this project, or the run references no such attachment.",
            "headers": {
              "X-Request-Id": {
                "description": "The id this server gave the request; quote it when reporting a problem.",
                "schema": {
                  "type": "string",
                  "format": "uuid"
                }
              }
            },
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "apiKey": {
        "type": "http",
        "scheme": "bearer",
        "description": "A project-scoped API key, `qer_k1_<publicId>_<secret>`, issued by an operator. Its scope (runs:read or runs:write) is named by each operation in x-qe-report-scope."
      }
    },
    "schemas": {
      "Problem": {
        "type": "object",
        "description": "RFC 9457 problem details. Further members depend on the code: `diagnostics` and `runId` for a refused run, `totalRuns`, `missingRuns`, and `staleRuns` for an incomplete query index.",
        "properties": {
          "type": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "status": {
            "type": "integer"
          },
          "code": {
            "type": "string",
            "enum": [
              "AUTHENTICATION_REQUIRED",
              "FORBIDDEN",
              "BAD_REQUEST",
              "NOT_FOUND",
              "METHOD_NOT_ALLOWED",
              "RUN_CONFLICT",
              "PAYLOAD_TOO_LARGE",
              "UNSUPPORTED_MEDIA_TYPE",
              "RUN_INVALID",
              "RUN_INCOMPLETE",
              "RUN_EMPTY",
              "INTERNAL_ERROR",
              "QUERY_INDEX_INCOMPLETE",
              "NOT_READY"
            ]
          },
          "detail": {
            "type": "string"
          },
          "requestId": {
            "type": "string"
          },
          "runId": {
            "type": "string"
          },
          "diagnostics": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Diagnostic"
            }
          },
          "totalRuns": {
            "type": "integer"
          },
          "missingRuns": {
            "type": "integer"
          },
          "staleRuns": {
            "type": "integer"
          },
          "problems": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "type",
          "title",
          "status",
          "code",
          "detail",
          "requestId"
        ],
        "additionalProperties": false
      },
...
      "IngestionResult": {
        "type": "object",
        "properties": {
          "outcome": {
            "type": "string",
            "enum": [
              "inserted",
              "already_present"
            ]
          },
          "runId": {
            "type": "string",
            "pattern": "^[\\x21-\\x7E]{1,128}$"
          },
          "runRef": {
            "type": "string",
            "pattern": "^[A-Za-z0-9_-]{2,172}$",
            "description": "Unpadded base64url of the UTF-8 run id: a transport locator only, never an identity."
          },
          "ingestionSequence": {
            "type": "string",
            "pattern": "^(0|[1-9][0-9]*)$",
            "description": "A decimal integer, as text: it may exceed what a JSON number holds exactly."
          },
          "blobRelationsAdded": {
            "type": "integer",
            "minimum": 0
          },
          "retentionAdded": {
            "type": "boolean"
          },
          "queryIndexRebuilt": {
            "type": "boolean"
          }
        },
        "required": [
          "outcome",
          "runId",
          "runRef",
          "ingestionSequence"
        ],
        "additionalProperties": false
      },
...
    }
  }
}